+++ Calculadora.h
@@ -2,7 +2,7 @@
 
 int soma(int operadorA, int operadorB);
 int subtracao(int minuendo, int subtraendo);
-int divisao(int numerador, int denominador);
+float divisao(int numerador, int denominador);
 int multiplicacao(int operadorA, int operadorB);
 int potenciacao(int base, int potencia);
